requirements.txt
accelerate
torch
torchvision
numpy
opencv-python
pillow
transformers
scipy
diffusers
requests
boto3
pillow
cloudscraper